move getFilterStatus to songs.js and update tests
