+++ apps/src/dance/DanceVisualizationColumn.jsx
@@ -8,8 +8,9 @@ import PropTypes from 'prop-types';
 import Radium from 'radium'; // eslint-disable-line no-restricted-imports
 import {connect} from 'react-redux';
 import i18n from '@cdo/locale';
-import AgeDialog, {getFilterStatus} from '../templates/AgeDialog';
+import AgeDialog from '../templates/AgeDialog';
 import {getFilteredSongKeys} from '@cdo/apps/dance/songs';
+import {getFilterStatus} from './songs';
 
 const SongSelector = Radium(
   class extends React.Component {
